requirements.txt
openai==1.30.1
python-dotenv==1.0.1
requests==2.32.2
PyYAML==6.0.1
